modification du type long en integer

+++ src/main/java/bat/tech/bd/models/Activity.java
@@ -9,7 +9,7 @@ public class Activity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idactivity;
+    private Integer idactivity;
     private Date date;
