+++ src/main/java/bat/tech/bd/models/Groupe.java
@@ -7,7 +7,7 @@ public class Groupe {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private Long idgroup;
+    private Integer idgroup;
     private String nomGroupe;
 
     @ManyToOne
